# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dtorkon API",
    "description": "FastAPI backend for the dtorkon mini-blog. The service handles admin-only auth, public/admin posts, site profile editing, analytics, SQLite persistence, searchable publishing, and backend-mediated uploads to Yandex Object Storage over the S3 API.",
    "version": "0.1.0-fastapi"
  },
  "servers": [
    {
      "url": "https://{host}",
      "variables": {
        "host": {
          "default": "example.com"
        }
      }
    }
  ],
  "paths": {
...
    "/api/site-profile": {
      "get": {
        "tags": [
          "Site Profile"
        ],
        "summary": "Get the public site profile",
        "operationId": "get_site_profile_api_site_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteProfileRead"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts": {
      "get": {
        "tags": [
          "Public Posts"
        ],
        "summary": "Get a paginated list of published posts",
        "operationId": "list_public_posts_api_posts_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 12,
              "title": "Page Size"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPostListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{slug}": {
      "get": {
        "tags": [
          "Public Posts"
        ],
        "summary": "Get a published post by slug",
        "operationId": "get_public_post_api_posts__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPostDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/site-profile": {
      "get": {
        "tags": [
          "Admin Site Profile"
        ],
        "summary": "Get the editable site profile",
        "operationId": "get_admin_site_profile_api_admin_site_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteProfileRead"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Admin Site Profile"
        ],
        "summary": "Update the editable site profile",
        "operationId": "update_admin_site_profile_api_admin_site_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSiteProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteProfileRead"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "mimeType": {
            "type": "string",
            "title": "Mimetype"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "originalName": {
            "type": "string",
            "title": "Originalname"
          },
          "status": {
            "$ref": "#/components/schemas/AssetStatus"
          },
          "transcriptStatus": {
            "$ref": "#/components/schemas/TranscriptStatus"
          },
          "transcriptText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcripttext"
          },
          "transcriptError": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcripterror"
          },
          "transcribedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcribedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "url",
          "mimeType",
          "size",
          "originalName",
          "status",
          "transcriptStatus"
        ],
        "title": "AssetRead"
      },
      "AssetStatus": {
        "type": "string",
        "enum": [
          "pending",
          "ready",
          "orphaned"
        ],
        "title": "AssetStatus"
      },
...
      "AttachmentKind": {
        "type": "string",
        "enum": [
          "image",
          "audio",
          "video",
          "file"
        ],
        "title": "AttachmentKind"
      },
      "AttachmentRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "assetId": {
            "type": "string",
            "title": "Assetid"
          },
          "kind": {
            "$ref": "#/components/schemas/AttachmentKind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "sortOrder": {
            "type": "integer",
            "title": "Sortorder"
          },
          "asset": {
            "$ref": "#/components/schemas/AssetRead"
          }
        },
        "type": "object",
        "required": [
          "id",
          "assetId",
          "kind",
          "title",
          "sortOrder",
          "asset"
        ],
        "title": "AttachmentRead"
      },
...
      "ErrorPayload": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorPayload"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Pagination": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize"
          },
          "totalItems": {
            "type": "integer",
            "title": "Totalitems"
          },
          "totalPages": {
            "type": "integer",
            "title": "Totalpages"
          }
        },
        "type": "object",
        "required": [
          "page",
          "pageSize",
          "totalItems",
          "totalPages"
        ],
        "title": "Pagination"
      },
      "PostStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published"
        ],
        "title": "PostStatus"
      },
...
      "PublicPostDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "bodyMarkdown": {
            "type": "string",
            "title": "Bodymarkdown"
          },
          "status": {
            "$ref": "#/components/schemas/PostStatus"
          },
          "coverAsset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentRead"
            },
            "type": "array",
            "title": "Attachments"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          },
          "publishedAt": {
            "type": "string",
            "title": "Publishedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "excerpt",
          "bodyMarkdown",
          "status",
          "attachments",
          "createdAt",
          "updatedAt",
          "publishedAt"
        ],
        "title": "PublicPostDetail"
      },
      "PublicPostListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "coverAsset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "publishedAt": {
            "type": "string",
            "title": "Publishedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "excerpt",
          "publishedAt"
        ],
        "title": "PublicPostListItem"
      },
      "PublicPostListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicPostListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "PublicPostListResponse"
      },
...
      "SiteProfileLinkInput": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/SiteProfileLinkKind"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": ""
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "url"
        ],
        "title": "SiteProfileLinkInput"
      },
      "SiteProfileLinkKind": {
        "type": "string",
        "enum": [
          "email",
          "phone",
          "telegram",
          "vk",
          "link"
        ],
        "title": "SiteProfileLinkKind"
      },
      "SiteProfileLinkRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/SiteProfileLinkKind"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "sortOrder": {
            "type": "integer",
            "title": "Sortorder"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "label",
          "url",
          "sortOrder"
        ],
        "title": "SiteProfileLinkRead"
      },
      "SiteProfileRead": {
        "properties": {
          "siteTitle": {
            "type": "string",
            "title": "Sitetitle"
          },
          "siteTagline": {
            "type": "string",
            "title": "Sitetagline"
          },
          "authorName": {
            "type": "string",
            "title": "Authorname"
          },
          "authorBio": {
            "type": "string",
            "title": "Authorbio"
          },
          "contactEmail": {
            "type": "string",
            "title": "Contactemail"
          },
          "avatarAssetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarassetid"
          },
          "avatarAsset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "backgroundColor": {
            "type": "string",
            "title": "Backgroundcolor"
          },
          "backgroundAssetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backgroundassetid"
          },
          "backgroundAsset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/SiteProfileLinkRead"
            },
            "type": "array",
            "title": "Links"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "siteTitle",
          "siteTagline",
          "authorName",
          "authorBio",
          "contactEmail",
          "backgroundColor",
          "updatedAt"
        ],
        "title": "SiteProfileRead"
      },
      "TranscriptStatus": {
        "type": "string",
        "enum": [
          "idle",
          "processing",
          "ready",
          "failed"
        ],
        "title": "TranscriptStatus"
      },
...
      "UpdateSiteProfileRequest": {
        "properties": {
          "siteTitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sitetitle"
          },
          "siteTagline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sitetagline"
          },
          "authorName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorname"
          },
          "authorBio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorbio"
          },
          "contactEmail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contactemail"
          },
          "avatarAssetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarassetid"
          },
          "backgroundColor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backgroundcolor"
          },
          "backgroundAssetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backgroundassetid"
          },
          "links": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SiteProfileLinkInput"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Links"
          }
        },
        "type": "object",
        "title": "UpdateSiteProfileRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dtorkon_session"
      }
    }
  },
...
}